pyproject.toml
[project]
name = "fpt-ai-residency"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-aiplatform>=1.86.0",
    "groq>=0.20.0",
    "jupyterlab>=4.3.6",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "openai>=1.69.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
